disabled diamonds refactoring
