+++ Workspace/DiamondKill/src/com/github/calfur/minecraftserverplugins/diamondkill/disabling/FeatureDisabler.java
@@ -10,7 +10,8 @@ public class FeatureDisabler {
 	private void registrateEvents() {
 		Main plugin = Main.getInstance();
 		plugin.getServer().getPluginManager().registerEvents(new EnchantingDisabler(), plugin);
-		plugin.getServer().getPluginManager().registerEvents(new NaturalDiamondDisabler(), plugin);
+		plugin.getServer().getPluginManager().registerEvents(new DiamondOreDisabler(), plugin);
+		plugin.getServer().getPluginManager().registerEvents(new ChestLootNerf(), plugin);
 		plugin.getServer().getPluginManager().registerEvents(new AnvilDisabler(), plugin);
 		plugin.getServer().getPluginManager().registerEvents(new VillagerTradesDisabler(), plugin);
 		plugin.getServer().getPluginManager().registerEvents(new PotionDisabler(), plugin);
